chore: add PKGCHECK_DEBUG_SDIST=1 support working on generated files
I've just been shoving a return in there, and one of these days I
was going to forget to remove it.  Thus do this instead.

+++ py_build.py
@@ -115,6 +115,11 @@ def create_generated_files():
         write_files(cleanup_files)
         yield
     finally:
+        if os.environ.get("PKGCHECK_DEBUG_SDIST"):
+            print(
+                "WARNING: not wiping cleanup files since PKGCHECK_DEBUG_SDIST is set in the environment"
+            )
+            return
         for path in cleanup_files:
             try:
                 path.unlink()
